requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Multi-agent and LLM
langchain>=0.1.0
langgraph>=0.0.20
langsmith>=0.1.0
openai>=1.3.0
chromadb>=0.4.0
langchain-openai>=0.1.0

# Memory and storage
mem0ai>=0.0.20
supabase>=2.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0
beautifulsoup4>=4.9.0

# Report generation
reportlab>=4.0.0
openpyxl>=3.1.0
markdown>=3.5.0

# Monitoring and logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Jupyter and development
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Optional: Add more dependencies as needed
# torch>=1.9.0
# tensorflow>=2.6.0
